// apps/pro-web/convex/_generated/api.d.ts
declare const fullApi: ApiFromModules<{
  action: typeof action;
  "action/details": typeof action_details;
  babel: typeof babel;
  betterAuthTriggers: typeof betterAuthTriggers;
  components: typeof components_;
  drafts: typeof drafts;
  http: typeof http;
  magicRock: typeof magicRock;
  migrations: typeof migrations;
  reactor: typeof reactor;
  "schedule/lifecycle": typeof schedule_lifecycle;
  schedules: typeof schedules;
  seed: typeof seed;
  skills: typeof skills;
  subscriptions: typeof subscriptions;
  tasks: typeof tasks;
  topUps: typeof topUps;
  transactions: typeof transactions;
  users: typeof users;
  "users/preferences": typeof users_preferences;
  "users/requests": typeof users_requests;
  "users/themes": typeof users_themes;
}>;
// ...
export declare const internal: FilterApi<
  typeof fullApi,
  FunctionReference<any, "internal">
>;